+++ erupt/src/utils/loading.rs
@@ -1,16 +1,13 @@
 //! Load functions using [`libloading`](https://crates.io/crates/libloading)
 //!
 //! Enabled using the `loading` cargo feature
-use crate::{EntryEnabled, EntryLoader, LoaderError};
+use crate::{CustomEntryLoader, EntryEnabled, LoaderError};
 use libloading::Library;
 use std::{
     error::Error,
     ffi::{CStr, OsStr},
     fmt::{self, Display},
 };
-
-/// The default `EntryLoader`, providing `EntryLoader::new`
-pub type DefaultEntryLoader = EntryLoader<Library>;
 
 #[cfg(all(unix, not(any(target_os = "macos", target_os = "ios", target_os = "android"))))]
 const LIB_PATH: &str = "libvulkan.so.1";
@@ -24,7 +21,13 @@ const LIB_PATH: &str = "libvulkan.dylib";
 #[cfg(windows)]
 const LIB_PATH: &str = "vulkan-1.dll";
 
-/// An error that can occur while loading a `Library`
+/// The default entry loader library, used in [`EntryLoader`]
+pub type DefaultEntryLoaderLibrary = Library;
+
+/// The default [`EntryLoader`], providing `EntryLoader::new`
+pub type EntryLoader = CustomEntryLoader<DefaultEntryLoaderLibrary>;
+
+/// An error that can occur while loading a library
 #[derive(Debug)]
 pub struct LibraryError(libloading::Error);
 
@@ -40,7 +43,7 @@ impl Error for LibraryError {
     }
 }
 
-/// An error that can occur while initializing a `EntryLoader`
+/// An error that can occur while initializing a [`EntryLoader`]
 #[derive(Debug)]
 pub enum EntryLoaderError {
     /// The library failed to load
@@ -69,19 +72,19 @@ impl Error for EntryLoaderError {
     }
 }
 
-impl DefaultEntryLoader {
+impl EntryLoader {
     /// Load functions using [`libloading`](https://crates.io/crates/libloading)
     ///
     /// Enabled using the `loading` cargo feature
-    pub fn new() -> Result<DefaultEntryLoader, EntryLoaderError> {
-        DefaultEntryLoader::with_lib_path(LIB_PATH)
+    pub fn new() -> Result<EntryLoader, EntryLoaderError> {
+        EntryLoader::with_lib_path(LIB_PATH)
     }
 
     /// Load functions using [`libloading`](https://crates.io/crates/libloading)
     /// providing a custom library path
     ///
     /// Enabled using the `loading` cargo feature
-    pub fn with_lib_path<P: AsRef<OsStr>>(lib_path: P) -> Result<DefaultEntryLoader, EntryLoaderError> {
+    pub fn with_lib_path<P: AsRef<OsStr>>(lib_path: P) -> Result<EntryLoader, EntryLoaderError> {
         let mut library = unsafe { Library::new(lib_path).map_err(|err| EntryLoaderError::Library(LibraryError(err)))? };
 
         let symbol = |library: &mut Library, name| unsafe {