+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added versioning warning to provisional/beta extensions
-- Allow for custom instance and device creation functions (useful for e.g. OpenXR)
+- Allow for more ergonomic custom instance and device creation functions (useful for e.g. OpenXR)
 
 ### Changed
 
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Split `ExtendableFrom` into `Const` and `Mut` variants
 - Remove function-specific "function not loaded" error messages
 - Expose `FramebufferCreateInfoBuilder::attachment_count`
+- Restructure `EntryLoader` naming
+  - `DefaultEntryLoader` -> `EntryLoader`
+  - `EntryLoader` -> `CustomEntryLoader`
 
 ### Fixed